changed line in reload method to self.all()[key] = classes.get(val['__class__'], BaseModel)(**val)
Okay so i made changes! parts got happy! Other parts got sad! Console.py is sad! but i think issue starts here with file_storage.py . This line attempts to access the __class__ attribute of the loaded object val to determine the class type. However, if __class__ is not present in the JSON data for some reason, it will result in a KeyError. Hoping that if val['__class__'] is not found in the classes dictionary, it defaults to BaseModel. If this doest work gonna look to change do_show in console.py with for stored_obj in storage.all().values():